+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # bandlab-android-intellij-plugin Changelog
 
 ## Unreleased
+### Added
+- Add Composable template action
 
 ## [2.4.1] - 2026-06-03
 ### Changed